Add security to product management to allow only superusers access.

+++ products/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render, get_object_or_404, redirect, reverse
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from .models import Product
 from .forms import ProductForm
 
@@ -28,8 +29,13 @@ def product_details(request, product_id):
     return render(request, 'products/product_details.html', context)
 
 
+@login_required
 def add_product(request):
     """ A view for admin to add products """
+    if not request.user.is_superuser:
+        messages.error(request, 'You do not have access to this function.')
+        return redirect(reverse('home'))
+
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
         if form.is_valid():
@@ -49,8 +55,13 @@ def add_product(request):
     return render(request, template, context)
 
 
+@login_required
 def edit_product(request, product_id):
     """ A view to edit the product """
+    if not request.user.is_superuser:
+        messages.error(request, 'You do not have access to this function.')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     
     if request.method == 'POST':
@@ -74,8 +85,13 @@ def edit_product(request, product_id):
     return render(request, template, context)
 
 
+@login_required
 def delete_product(request, product_id):
     """ A view to delete products """
+
+    if not request.user.is_superuser:
+        messages.error(request, 'You do not have access to this function.')
+        return redirect(reverse('home'))
 
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
